+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ENHANCEMENTS
 
-* resource/aws_kinesis_firehose_delivery_stream: Add `http_endpoint` configuration block [GH-15356]
+* resource/aws_kinesis_firehose_delivery_stream: Add `http_endpoint_configuration` configuration block [GH-15356]
 
 ## 3.14.1 (November 06, 2020)